+++ src/app/modules/lux-form/lux-autocomplete/lux-autocomplete.component.ts
@@ -1,7 +1,8 @@
 import {
   AfterViewInit,
   ChangeDetectorRef,
-  Component, ContentChild,
+  Component,
+  ContentChild,
   ElementRef,
   EventEmitter,
   Input,
@@ -12,7 +13,7 @@ import {
   ViewChild
 } from '@angular/core';
 import { ControlContainer } from '@angular/forms';
-import { MatAutocompleteSelectedEvent, MatAutocompleteTrigger } from '@angular/material/autocomplete';
+import { MatAutocomplete, MatAutocompleteSelectedEvent, MatAutocompleteTrigger } from '@angular/material/autocomplete';
 import { LuxFormComponentBase, LuxValidationErrors } from '../lux-form-model/lux-form-component-base.class';
 import { debounceTime, distinctUntilChanged, map, startWith } from 'rxjs/operators';
 import { ReplaySubject, Subscription } from 'rxjs';
@@ -20,7 +21,6 @@ import { LuxConsoleService } from '../../lux-util/lux-console.service';
 import { LuxComponentsConfigService } from '../../lux-components-config/lux-components-config.service';
 import { LuxInputPrefixComponent } from '../lux-input/lux-input-subcomponents/lux-input-prefix.component';
 import { LuxInputSuffixComponent } from '../lux-input/lux-input-subcomponents/lux-input-suffix.component';
-
 
 @Component({
   selector: 'lux-autocomplete',
@@ -33,12 +33,16 @@ export class LuxAutocompleteComponent<V = any, O = any> extends LuxFormComponent
   private valueChangeSubscription?: Subscription;
 
   filteredOptions: O[] = [];
+  displayedOptions: O[] = [];
+  loadingRunning = false;
+  activeIndex = -1;
   _luxOptions: O[] = [];
 
   @Input() luxPlaceholder = '';
   @Input() luxOptionLabelProp = 'label';
   @Input() luxLookupDelay = 500;
-  @Input() luxErrorMessageNotAnOption = $localize`:@@luxc.autocomplete.error_message.not_an_option:Der eingegebene Wert ist nicht Teil der Auswahl.`;
+  @Input()
+  luxErrorMessageNotAnOption = $localize`:@@luxc.autocomplete.error_message.not_an_option:Der eingegebene Wert ist nicht Teil der Auswahl.`;
   @Input() luxTagId?: string;
   @Input() luxSelectAllOnClick = true;
   @Input() luxStrict = true;
@@ -50,6 +54,7 @@ export class LuxAutocompleteComponent<V = any, O = any> extends LuxFormComponent
   @Input() luxNoTopLabel = false;
   @Input() luxNoBottomLabel = false;
   @Input() luxOptionMultiline = false;
+  @Input() luxOptionBlockSize = 500;
 
   @Output() luxValueChange = new EventEmitter<V | null>();
   @Output() luxOptionSelected = new EventEmitter<V | null>();
@@ -61,6 +66,7 @@ export class LuxAutocompleteComponent<V = any, O = any> extends LuxFormComponent
 
   @ViewChild('autoCompleteInput', { read: MatAutocompleteTrigger }) matAutoComplete!: MatAutocompleteTrigger;
   @ViewChild('autoCompleteInput', { read: ElementRef }) matInput!: ElementRef;
+  @ViewChild(MatAutocomplete) matAutocompleteComponent!: MatAutocomplete;
 
   get luxValue(): V {
     return this.getValue();
@@ -143,7 +149,71 @@ export class LuxAutocompleteComponent<V = any, O = any> extends LuxFormComponent
       })
     );
 
+    this.subscriptions.push(
+      this.matAutocompleteComponent._keyManager.change.subscribe((index) => {
+        if (this.loadingRunning && index === -1) {
+          // Workaround: Bei Änderungen an den Optionen wird der Aktivindex zurückgesetzt!
+          //
+          // Beim Nachladen werden die Optionen verändert und der Aktivindex
+          // im KeyManager wird zurückgesetzt. D.h. der nächste Klick auf die
+          // Pfeiltaste (nach unten) aktiviert nicht die nächste Option, sondern
+          // die erste Option am Anfang der Liste. Aus diesem Grund wird hier
+          // der letzte Aktivindex wiederhergestellt, damit der Benutzer dort
+          // weitermachen kann, wo er aufgehört hat.
+          //
+          // Siehe: _MatAutocompleteTriggerBase._subscribeToClosingActions
+          // this._resetActiveItem();
+          setTimeout(() => {
+            this.matAutocompleteComponent._keyManager.setActiveItem(this.activeIndex!);
+            this.loadingRunning = false;
+          });
+        }
+      })
+    );
+
+    this.subscriptions.push(
+      this.matAutocompleteComponent.opened.subscribe(() => {
+        setTimeout(() => {
+          if (this.matAutocompleteComponent.panel) {
+            this.matAutocompleteComponent.panel.nativeElement.addEventListener('scroll', this.loadOnScroll.bind(this));
+          }
+        });
+      })
+    );
+
+    this.subscriptions.push(
+      this.matAutocompleteComponent.closed.subscribe(() => {
+        this.updateFilterOptions();
+        this.matAutocompleteComponent.panel.nativeElement.removeEventListener('scroll', this.loadOnScroll);
+      })
+    );
+
     this.registerNewValueChangesListener();
+  }
+
+  /**
+   * Stößt das Nachladen von Elementen an, wenn ein bestimmter Scrollwert erreicht wurde.
+   *
+   * @param event - ScrollEvent
+   */
+  private loadOnScroll(event: Event) {
+    const position = event.target as any;
+    if (position && (position.scrollTop + position.clientHeight) / position.scrollHeight > 85 / 100) {
+      this.updateDisplayedEntries();
+    }
+  }
+
+  /**
+   * Läd den nächsten Block Daten aus den Entries nach.
+   */
+  updateDisplayedEntries() {
+    if (this.filteredOptions.length > 0) {
+      this.loadingRunning = true;
+      this.activeIndex = this.matAutocompleteComponent._keyManager.activeItemIndex ?? -1;
+      const start = 0;
+      const end = Math.min(this.luxOptionBlockSize, this.filteredOptions.length);
+      this.displayedOptions.push(...this.filteredOptions.splice(start, end));
+    }
   }
 
   /**
@@ -241,7 +311,7 @@ export class LuxAutocompleteComponent<V = any, O = any> extends LuxFormComponent
   private handleErrors() {
     const errors = this.formControl ? this.formControl.errors : null;
     if (
-      this.luxOptions.indexOf(this.luxStrict ? this.getPickValueOption(this.luxValue as any) : this.luxValue as any) > -1 ||
+      this.luxOptions.indexOf(this.luxStrict ? this.getPickValueOption(this.luxValue as any) : (this.luxValue as any)) > -1 ||
       (!!errors && Object.keys(errors).length > 0 && errors['required'])
     ) {
       this.handleOtherErrors(errors);
@@ -271,7 +341,7 @@ export class LuxAutocompleteComponent<V = any, O = any> extends LuxFormComponent
   notifyFormValueChanged(formValue: any) {
     let newValue;
     if (this.luxStrict) {
-      newValue = formValue instanceof Object && !!this.luxPickValue? this.luxPickValue(formValue) : formValue;
+      newValue = formValue instanceof Object && !!this.luxPickValue ? this.luxPickValue(formValue) : formValue;
     } else {
       newValue = formValue;
     }
@@ -284,9 +354,9 @@ export class LuxAutocompleteComponent<V = any, O = any> extends LuxFormComponent
   }
 
   private getPickValueOption(value: O): O | null {
-    const pickValue = value instanceof Object && !!this.luxPickValue? this.luxPickValue(value) : value;
+    const pickValue = value instanceof Object && !!this.luxPickValue ? this.luxPickValue(value) : value;
     const found = this.luxOptions.find((currentOption) => {
-      const pickOptionValue = currentOption instanceof Object && !!this.luxPickValue? this.luxPickValue(currentOption) : currentOption;
+      const pickOptionValue = currentOption instanceof Object && !!this.luxPickValue ? this.luxPickValue(currentOption) : currentOption;
       return pickValue === pickOptionValue;
     });
 
@@ -294,7 +364,9 @@ export class LuxAutocompleteComponent<V = any, O = any> extends LuxFormComponent
   }
 
   private updateFilterOptions() {
-    this.filteredOptions =  this.filterOptions();
+    this.filteredOptions = this.filterOptions();
+    this.displayedOptions = [];
+    this.updateDisplayedEntries();
   }
 
   private registerNewValueChangesListener() {
@@ -304,23 +376,24 @@ export class LuxAutocompleteComponent<V = any, O = any> extends LuxFormComponent
     }
 
     // Eine neue Subscription hinzufügen.
-    this.valueChangeSubscription =
-      this.formControl.valueChanges
-        .pipe(
-          startWith(""),
-          debounceTime(this.luxLookupDelay),
-          map(() => {
-            return this.filterOptions();
-          })
-        )
-        .subscribe((result) => {
-          this.filteredOptions = result;
-        });
+    this.valueChangeSubscription = this.formControl.valueChanges
+      .pipe(
+        startWith(''),
+        debounceTime(this.luxLookupDelay),
+        map(() => {
+          return this.filterOptions();
+        })
+      )
+      .subscribe((result) => {
+        this.filteredOptions = result;
+        this.displayedOptions = [];
+        this.updateDisplayedEntries();
+      });
   }
 
   private filterOptions() {
     const filterLabel = this.matInput.nativeElement.value;
-    return filterLabel ? this.filter(filterLabel) : this.luxOptions;
+    return filterLabel ? this.filter(filterLabel) : [...this.luxOptions];
   }
 
   private updateFormControlValue() {
@@ -330,9 +403,9 @@ export class LuxAutocompleteComponent<V = any, O = any> extends LuxFormComponent
       if (filterResult.length === 1) {
         let selected;
         if (this.luxStrict && !!this.luxPickValue) {
-          selected = this.luxPickValue(filterResult[ 0 ]);
+          selected = this.luxPickValue(filterResult[0]);
         } else {
-          selected = filterResult[ 0 ];
+          selected = filterResult[0];
         }
         this.formControl.setValue(selected as any);
       }