+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## New
 
+- **allgemein**: Update auf LUX-Components-Theme [14.4.0](https://github.com/IHK-GfI/lux-components-theme/releases/tag/14.4.0).
 - **allgemein**: Update auf LUX-Components-Icons-And-Fonts. [1.4.0](https://github.com/IHK-GfI/lux-components-icons-and-fonts/releases/tag/1.4.0).
 - **styles**: Webfonts werden über das Theme geladen. [Issue 277](https://github.com/IHK-GfI/lux-components/issues/277)
 - **lux-autocomplete**: Performance mit großen Datenmengen verbessern. [Issue 278](https://github.com/IHK-GfI/lux-components/issues/278)